Improve t9 converter code organization. Add options to converter
